loader: improve fsnotify error message (#29)
loader: improve fsnotify error message and use Panic() instead of Fatal()

Use different error messages for failing to create a watcher and
watching a file.  Additionally, use Panicf() instead of Fatalf()
otherwise it is impossible to identify the calling function.

Note: this library *really* should not be be calling Fatal() or 
Panic(), but should instead be returning an error.  But changing
that is outside the scope of this change as it would change the 
package API.

Additionally, fix calls to Warnf() in loader