loader: add New2 constructor that returns an error. (#31)
* loader: add New2 constructor that returns an error.

The existing loader.New constructor panics on error - this is not
acceptable in libraries.  This commit adds New2 which returns an
error instead of panicking.

* loader: fix tests by using Rename() instead of writing the file

Basically, there is a race were the loader detects a change and scans
the parent directory before we are done writing the new file.  This
fixes that by using rename.