snapshot: use per-Snapshot PRNGs to reduce lock contention + cleanup
This commit changes the snapshot package to use per-Snapshot PRNGs
instead of a single global PRNG - this reduces lock contention.

Additionally, it cleans up the code a bit by changing map accesses to
check if the result if nil instead of doing the "val, ok" check.

These changes are pretty minor and are mostly a nit.